// web/vite-graphql/src/components/__generated__/BookDetails_book.graphql.ts
export type BookDetails_book$data = {
  readonly characters: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly name: string;
      } | null | undefined;
    }> | null | undefined;
    readonly pageInfo: {
      readonly hasNextPage: boolean;
    };
  } | null | undefined;
  readonly id: string;
  readonly title: string;
  readonly " $fragmentType": "BookDetails_book";
};
export type BookDetails_book$key = {
  readonly " $data"?: BookDetails_book$data;
  readonly " $fragmentSpreads": FragmentRefs<"BookDetails_book">;
};